Completed: Finished Magic-8-ball application
This is a tutorial application to learn flutter. This is an application part of the flutter course on The App Brewery(https://www.appbrewery.co). This application will allow user to ask their questions and then tap on the middle of the ball in the UI to get a random answer that are statically fed in this application.